+++ src/full/Agda/Termination/Monad.hs
@@ -545,6 +545,8 @@ instance UsableSizeVars DeBruijnPattern where
     LitP{}     -> none
     DotP{}     -> none
     ProjP{}    -> none
+    IApplyP{}  -> none
+    DefP{} -> none
     where none _ = return mempty
 
 instance UsableSizeVars [DeBruijnPattern] where
@@ -562,6 +564,8 @@ instance UsableSizeVars (Masked DeBruijnPattern) where
     LitP{}     -> none
     DotP{}     -> none
     ProjP{}    -> none
+    IApplyP{}  -> none
+    DefP{}     -> none
     where none _ = return mempty
 
 instance UsableSizeVars MaskedDeBruijnPatterns where