+++ src/full/Agda/TypeChecking/CompiledClause/Compile.hs
@@ -172,8 +172,10 @@ compile cs = case nextSplit cs of
     name (VarP _ x) = x
     name (DotP _ _) = underscore
     name ConP{}  = __IMPOSSIBLE__
+    name DefP{}  = __IMPOSSIBLE__
     name LitP{}  = __IMPOSSIBLE__
     name ProjP{} = __IMPOSSIBLE__
+    name (IApplyP _ _ _ x) = x
 
 -- | Get the index of the next argument we need to split on.
 --   This the number of the first pattern that does a (non-lazy) match in the first clause.
@@ -199,8 +201,10 @@ nextSplit (Cl ps _ : cs) = findSplit nonLazy ps <|> findSplit allAgree ps
 --   And if yes, is it a record pattern?
 properSplit :: Pattern' a -> Maybe Bool
 properSplit (ConP _ cpi _) = Just (Just PatORec == conPRecord cpi)
+properSplit DefP{} = Just False
 properSplit LitP{}  = Just False
 properSplit ProjP{} = Just False
+properSplit IApplyP{} = Nothing
 properSplit VarP{}  = Nothing
 properSplit DotP{}  = Nothing
 
@@ -211,8 +215,10 @@ isVar :: Pattern' a -> Bool
 isVar VarP{}  = True
 isVar DotP{}  = True
 isVar ConP{}  = False
+isVar DefP{} = False
 isVar LitP{}  = False
 isVar ProjP{} = False
+isVar IApplyP{} = True
 
 -- | @splitOn single n cs@ will force expansion of catch-alls
 --   if @single@.
@@ -224,8 +230,11 @@ splitOn single n cs = mconcat $ map (fmap (:[]) . splitC n) $
 splitC :: Int -> Cl -> Case Cl
 splitC n (Cl ps b) = caseMaybe mp fallback $ \case
   ProjP _ d   -> projCase d $ Cl (ps0 ++ ps1) b
+  IApplyP{}   -> fallback
   ConP c i qs -> (conCase (conName c) (conPFallThrough i) $ WithArity (length qs) $
                    Cl (ps0 ++ map (fmap namedThing) qs ++ ps1) b) { lazyMatch = conPLazy i }
+  DefP o q qs -> (conCase q False $ WithArity (length qs) $
+                   Cl (ps0 ++ map (fmap namedThing) qs ++ ps1) b) { lazyMatch = False }
   LitP l      -> litCase l $ Cl (ps0 ++ ps1) b
   VarP{}      -> fallback
   DotP{}      -> fallback