+++ src/full/Agda/TypeChecking/Datatypes.hs
@@ -52,6 +52,17 @@ getConstructorData c = do
     Constructor{conData = d} -> return d
     _                        -> __IMPOSSIBLE__
 
+-- | Is the datatype of this constructor a Higher Inductive Type?
+--   Precondition: The argument must refer to a constructor of a datatype or record.
+consOfHIT :: QName -> TCM Bool
+consOfHIT c = do
+  d <- getConstructorData c
+  def <- theDef <$> getConstInfo d
+  case def of
+    Datatype {dataPathCons = xs} -> return $ not $ null xs
+    Record{} -> return False
+    _  -> __IMPOSSIBLE__
+
 -- | @getConType c t@ computes the constructor parameters from type @t@
 --   and returns them plus the instantiated type of constructor @c@.
 --   This works also if @t@ is a function type ending in a data/record type;