+++ src/full/Agda/TypeChecking/Errors.hs
@@ -1282,8 +1282,12 @@ instance PrettyTCM TypeError where
     prettyPat n (I.ConP c _ args) =
       mpar n args $
         prettyTCM c <+> fsep (map (prettyArg . fmap namedThing) args)
+    prettyPat n (I.DefP o q args) =
+      mpar n args $
+        prettyTCM q <+> fsep (map (prettyArg . fmap namedThing) args)
     prettyPat _ (I.LitP l) = prettyTCM l
     prettyPat _ (I.ProjP _ p) = "." <> prettyTCM p
+    prettyPat _ (I.IApplyP _ _ _ _) = "_"
 
 notCmp :: Comparison -> TCM Doc
 notCmp cmp = "!" <> prettyTCM cmp