+++ src/full/Agda/TypeChecking/Injectivity.hs
@@ -378,7 +378,9 @@ invertFunction cmp blk (Inv f blkArgs hdMap) hd fallback err success = do
 
     metaPat (DotP _ v)       = dotP v
     metaPat (VarP _ _)       = nextMeta
+    metaPat (IApplyP{})      = nextMeta
     metaPat (ConP c mt args) = Con c (fromConPatternInfo mt) . map Apply <$> metaArgs args
+    metaPat (DefP o q args)  = Def q . map Apply <$> metaArgs args
     metaPat (LitP l)         = return $ Lit l
     metaPat ProjP{}          = __IMPOSSIBLE__