+++ src/full/Agda/TypeChecking/MetaVars.hs
@@ -472,6 +472,13 @@ etaExpandMeta kinds m = whenM (asksTC envAssignMetas `and2M` isEtaExpandable kin
       IsSort{} -> dontExpand
       HasType _ a -> do
         TelV tel b <- telView a
+
+        -- Eta expanding metas with a domFinite will just make sure
+        -- they go unsolved: conversion will compare them at the
+        -- different cases for the domain, so it will not find the
+        -- solution for the whole meta.
+        if any domFinite (flattenTel tel) then dontExpand else do
+
         -- if the target type @b@ of @m@ is a meta variable @x@ itself
         -- (@NonBlocked (MetaV{})@),
         -- or it is blocked by a meta-variable @x@ (@Blocked@), we cannot
@@ -1071,7 +1078,7 @@ instance ReduceAndEtaContract Term where
       -- In case of lambda or record constructor, it makes sense to
       -- reduce further.
       Lam ai (Abs x b) -> etaLam ai x =<< reduceAndEtaContract b
-      Con c ci es | Just args <- allApplyElims es -> etaCon c ci args $ \ r c ci args -> do
+      Con c ci es -> etaCon c ci es $ \ r c ci args -> do
         args <- reduceAndEtaContract args
         etaContractRecord r c ci args
       v -> return v